refactor(ivc): move IVC aperture allocator and notify endpoint from axdevice to axvm
