+++ virtualization/axdevice/src/lib.rs
@@ -41,7 +41,6 @@ mod model;
 #[cfg(any(target_arch = "loongarch64", test))]
 #[cfg_attr(test, allow(dead_code))]
 mod loongarch_pch_pic;
-mod range_alloc;
 mod registration;
 mod resources;
 mod runtime_resources;
@@ -77,10 +76,6 @@ pub use loongarch_pch_pic::{
     PchPicOutputPort, PchPicOutputPortKey,
 };
 pub use model::{DeviceFirmwareProperty, DeviceFirmwareSpec, DeviceModel};
-pub use range_alloc::{
-    IvcApertureAllocator, IvcApertureAllocatorKey, IvcAperturePool, IvcNotifyEndpoint,
-    IvcNotifyIrqKey, WiredIvcNotifyEndpoint,
-};
 pub use registration::{
     DeviceBundle, DeviceLifecycle, DeviceRegistration, DmaPollableDeviceOps, PollableDeviceOps,
 };
