+++ virtualization/axvm/src/runtime/hvc.rs
@@ -665,13 +665,12 @@ impl HyperCall {
                         Ok(())
                     })
                     .map_err(|error| self.operation_error("wake IVC notify target VM", error))?;
-                let notify_irq = target_vm
-                    .get_devices()
-                    .map_err(|error| self.operation_error("get IVC notify target devices", error))?
-                    .notify_ivc_peer()
-                    .map_err(|error| {
-                        self.operation_error("notify IVC peer interrupt", error.into())
-                    })?;
+                let target_devices = target_vm.get_devices().map_err(|error| {
+                    self.operation_error("get IVC notify target devices", error)
+                })?;
+                let notify_irq = ivc::notify_peer(&target_devices).map_err(|error| {
+                    self.operation_error("notify IVC peer interrupt", error.into())
+                })?;
                 info!(
                     "IVC notify source VM[{}] target VM[{}] publisher VM[{}] key {:#x} irq={:?}",
                     route.source_vm_id,
